requirements.txt

Django==2.2.6
django-allauth==0.39.1
django-braces==1.13.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-model-utils==3.2.0
django-money==1.0
django-star-ratings==0.8.0
Pillow==6.0.0
reportlab==3.5.18
stripe==2.27.0
twilio==6.33.1

